+++ lib/async.dart
@@ -20,6 +20,7 @@ export "src/result_future.dart";
 export "src/stream_completer.dart";
 export "src/stream_group.dart";
 export "src/stream_queue.dart";
+export "src/stream_sink_transformer.dart";
 export "src/stream_splitter.dart";
 export "src/subscription_stream.dart";
 export "stream_zip.dart";